backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1
pydantic==2.7.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
bcrypt==4.1.3
python-multipart==0.0.9
redis==5.0.4
aioredis==2.0.1
slowapi==0.1.9
httpx==0.27.0
python-dotenv==1.0.1
structlog==24.1.0
psycopg2-binary==2.9.9
greenlet==3.0.3
email-validator==2.1.1
